requirements.txt
asgiref==3.8.1
astroid==3.3.10
attrs==25.3.0
dill==0.4.0
dj-database-url==2.3.0
Django==5.2.1
django-seed==0.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
Faker==37.3.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
tomlkit==0.13.2
toposort==1.10
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
